Minor bugfix and preparation for minor API rework
Bugfix: implemented destroying enemies in a tunnel by moving the pipe head to the enemy's position

Also added a new configuration file and some classes, enums and interfaces as preparation for reworking the API
to make the code more readable and less complex by representing the network of underground tunnels as a graph
structure implemented via linked node (tunnel cell) objects with pointers to nodes in all four directions (left,
right, up, down; a wall in this case is represented as a null value for that direction). It was already tried before,
but there were some bugs in the initial implementation, so we are going to try it again, but we need first of all
a working game API right now, so a more thorough reworking of the game API is postponed to a later date.