Added additional protection from cheats and reorganized code.
This commit builds on the previous commit, where we introduced a proxy game world object,
that is passed to the player instead of the actual game object to prevent the player
cheating by accessing game methods that we do not want him to access (registerActor(),
unregisterActor(), update(), etc.).

This commit introduces a further layer of protection from cheats by requiring the caller
of any "privileged" methods of other actors and game objects that are now publicly
accessible (ManipulableActor.act(), Bullet.launch(), TunnelCell.setAtDirection() etc.)
to authorize by passing a ManipulableGameInterface object (a game world object) that is
not null and at the same time is equal to the actual game world object stored in that actor
or object.

Since all objects and actors the student can't program do have this reference, whereas
the student only has a proxy object to the game, this effectively means that those methods,
though now visible to the player, have no effect on the game at all.

This also enables us to create a more hierarchical package structure for the game objects.
The previous strategy was trying to make all methods the student should not use package-local,
but this required using a flat package hierarchy with all game objects in one package, which
makes it harder to navigate in the project, and it didn't work all the time.