DRY'd and refactored code (related to actor movements)
Refactored and moved code related to moving actors to one place (AbstractActor class).

The changes in bullet points:
- Actors can return a reference to their current position in the form of a TunnelCell object.
It should be noted that this TunnelCell object's coordinates need not be equal to the x and y
coordinates of the actor:

the actor's x position is between cell.getX() and cell.getX() + world.getOffsetX()
the actor's y position is between cell.getY() anf cell.getY() + world.getOffsetY()

Or, to show it, imagine the dot in this picture represents the actor, with its x and y
coordinates:

               __________ (cellX + offsetX, cellY + offsetY)
              |          |
              |          |
              |  .(x,y)  |
              |          |
              |__________|
(cellX, cellY)

cellX = cell.getX();
cellY = cell.getY();
offsetX = world.getOffsetX();
offsety = world.getOffsetY();

- Moving an actor is now allowed only through the move() method in their AbstractActor superclass.
This method, besides the x and y coordinates, also updates the current position as well as direction
(orientation) of the actor, and also sets a boolean variable indicating whether there is no adjacent
TunnelCell in the given direction from the actors current position (i.e., a wall). This was primarily
done to extract common code from the Enemy and Bullet classes.
- The enum class now includes a method for getting a direction value based on the position shifts
in the x and y direction. The shift has to be 0 in exactly one direction for this method to return
a valid value, otherwise an exception is thrown.
- The PipeHead class now extends AbstractActor and has access to all its methods including move().
Together with the changes in the Direction enum class, this means all code responsible for updating
the head after push() and pop() is now simplified.
- The PipeSegment class now allows to specify the segment type indirectly by saying where the pipe
came from and where it is going. Once again, with the changes in the enum class, this helps simplify
much code.