Refactoring (still a lot to do) and new game ideas.
First attempts at structuring the game into distinct levels more focused on specific subtopics
to teach students (i.e. one level to teach stack operations, another more focused on queue
operations, another which includes enemies and an ultimate level featuring everything above).

These levels are represented by values of a new enum class GameType. Since each of these levels
requires additional code to be called on game update, we have decided to delegate the game world
initialization code to a builder class (TreasureScooperBuilder) and any updating related code
to a range of updater classes implementing a common interface (GameUpdater). The game world
stores these updater objects in a set which it iterates over in its update() method, calling
each elements update() method. By using the builder, we can set custom updaters for each level.

Speaking of the queue level, a gameplay idea we have come up with is to let walls randomly appear
in the game in the network of tunnels which block the path, an the only way to deal with them is
to use the pipe's weapon, operated as a queue. This might ultimately be a more viable idea than
the original idea of shooting randomly generated enemies, or a good alternative for teaching queue
operations, mainly because the pipe is not endangered, so no risk of losing, it just can't collect
everything in the underground tunnel network, so the player can't win either, unless he puts thought
into properly using the weapon.

An initial implementation of this concept is included, but it is still incomplete (walls can appear
even where other actors or the pipe is). As an aside note, the class hierarchy for actors has been
refactored, since we felt the wall is a kind of actor, mainly because it needs the same functionality,
and AbstractActor already has most of the functionality it needs.
Current hierarchy:

Enemy---------+
              |
              +---->AbstractMoveableActor-------+
              |                                 |
Bullet--------+                                 |
                                                +--->AbstractActor-----implements------><<Actor>>
                                                |
                                                |
                                     Wall-------+
TODO: optimize and refactor this new code