Wall implementation finalized, also some bugfixes.
The implementation of random wall generation as an alternative strategy
for teaching queue operations has been corrected to a working state,
though it will probably need some fixing later on.

Also rewrote the structure of the game interfaces to simplify their hierarchy.
The reason is that all enemy related functionality is now handled through
actor-related functions, which are kept even when there are no enemies
in a specific level, because bullets are also actors, and they are available
at every level, hence, there is no more reason to differetiate between
non-enemy-inclusive and enemy-inclusive levels of the game by creating multiple
interfaces.

Also minor bugfixes.