Minor bugfixes, optimization and documentation of game updaters
Removed the unnecessary game parameter from the definition of a gameupdater interface
as it was never used in the first place (except maybe as simple authentication) and
is not even necessary now.

Also reworked the implementation of updaters responsible for managing enemy and wall
creation and destruction, to simplify the code. Among changes:
-- they no longer maintain a list of actors they created, only a count of them,
   which is incremented on creation and decremented when the actor calls back its
   creating class.
-- also, they now support manually setting the number of iterations of the game loop
   before creating a new actor, as opposed to setting it to a random number (though
   the latter still is possible).

Also, all updaters have been documented with comments in their source code (probably
will be updated later).