Added multiple level support, multiplayer support pending.
The game now suports using multiple levels each of them with different
game worlds and different features per level (no enemies in one level,
walls present in another level, ...).

The changes done to enable such support:

A)

Created a GameManager class to handle the levels and switch between them.
The Game renderer now holds a reference to the game manager and renders
each level one by one while playing and a little time after the player
has won or lost in a particular level. It plays each level one by one,
and once the last level completes, the game shuts itself down.

B)

Since each level can feature a different tunnel maze or different world,
the rendering has been refactored. Now, the background is featureless,
and the tunnel maze is rendered upon it at each iteration, cell by cell.

C)

Related to point B), images for each tunnel cell type have been added,
and the background image changed.

D)

Also related to the use of different game worlds, the game configuration
written in XML has 2 layers to it. There is now a master configuration
file (config.xml) which contains information about each particular level
(and some additional info for the renderer). Each game world is configured
in its own file and these are referenced by attributes in the master config.
We have also included a DTD file to validate the configuration of each world,
although it does not guarantee correct content, only syntax. Also included
is documentation those world config files in the README.txt file.

As an aside note, the reason to separate this coniguration into 2 layers
was to enable reuse of a particular game world in multiple levels.

D)

As can be expected, the changes in the XML configuration also mean changes
to the processing of this configuration in the actual program code.
To be brief: There are now 2 XML configuration processors, one for the game
and one for the game worlds.

Also added the support for multiple entrances to the tunnel maze,
as a first step to implementing multiplayer scenarios.