Moved all game accessories and services to a separate class
Moved all features that the player can 'buy' into a separate GameShop class.
Specifically: buying bullets or repairing the pipe in exchange for score points.
This functionality used to be included in the Pipe class.

Also replaced the Pipe's damagePipe() and repairPipe() methods with setHealth()
as it can be used by the shop as well as Enemies etc. and added a callback
interface to change the score upon purchasing something from the game shop.