Changed actor removal functionality and added documentation to interfaces
Changed the way actors are removed from the game. Now, any code
that should be called has to be set when the actor is registered,
and is called upon unregistering the actor (previously, it was
rather incosistent).

Also added some small rendering effects (fading out the sprite of an actor
into the background) for actors that were unregistered.

Also, the actor manager is now responsible for updating all game actors.

Also added JavaDoc documentation to all interfaces.