Modified proxy objects interfaces and the code that uses their functions.
As the game world is immutable and that is not expected to change in the future,
there really is no reason for the player to be unable to access it and all its
attributes directly rather than via proxy object.

This of course means that anytime actors want to access values like offsetX, width,
a list of all tunnels etc., they have to access it through the GameWorld object,
rather than the proxy like before, so some code had to be changed there as well.