Modified pipe, tunnelcell and actor movement functionality
Pipe: a PipeSegment class is now a subclass of the base AbstractActor class.

TunnelCell: a TunnelCell object can now check if a given set of coordinates
(x, y) is within that tunnel cell. Also a unit test to test this method was
added.

Actor movement: The new TunnelCell functionality allowed for changing code
resposible for actor movement, somewhat simplifying it and making it a bit
more independent of the coordinate system used.

Also: The same (theoretical) independence from the coordinate system was
added to the code that is responsible for building the coordinate system.
The only invariant for TunnelCells and HorizontalTunnels (and implicitly
actors) is that their (x, y) coordinates represent the lower left edge
of their sprite or other graphical representation on the game screen
(HorizontalTunnel doesn't have a sprite, as it is a collection of cells).