Refactored actor rendering
The selection of the approriate sprite for an actor is now determined
by that actors class as opposed to the ActorType enum. This enables
multiple similar but distinct types of actors (for example, multiple
types of enemies) to be rendered with their appropriate sprite.

Also changed the config file and image assets to reflect this usage.
In the game config, the attribute that used to signify an ActorType
enum for an actor was now changed to reflect its class name. As all
actors are in the same package, it is simple to find its class.