Changed game level initialization
Updater classes can now be specified in the external config
and instantiated at the time a new level is built.

This uses basically the same scheme as the player instantiation
scheme introduced in commit 61c3cfa with multiplayer support.

Also, as the GameType enum basically was just to specify updaters,
and now is no longer needed, it was deleted.