Changed actor manager and game level purpose and interfaces.
Since a player can be considered a type of actor (not in the sense of rendering,
but as a subject making decisions within the game) the actor manager now handles
players as well as all other actors (also because the code looks nicer this way).

The game level now only provides indirect access to all players via actor manager.
The actor manager in turn returns a map of all players and the pipe they control.
Note that this map is not modifiable directly.

Because of the reworking of both sets of interfaces, the proxies for the player
and actors have been changed.

Also added a GameStateTester interface and its first implementation. Any implementation
of this interface shall be responsible for evaluating the current game state. This way,
if someone wants to use a different set of criteria for evaluating whether the player(s)
have won or lost the current game level, they only have to implement this interface and
use it instead of the default implementation.

TODO: One issue with multiplayer support is that currently, there is no mechanism ensuring
that one player cannot use the pipe of another player. Also, there remain some methods,
use of which by a player would be considered cheating (e.g. TunnelCell.setAtDirection())
and which do not yet have a specific mechanism of safekeeping their use (in other words,
no insurance the player can't use them). We already have an idea in mind, which we will
try to implement next.