Added method call authentication to prevent player cheats.
As mentioned in the previous commit (419d6ce) we have introduced a new authentication mechanism
to prevent players cheating by calling methods that they should not call.

The mechanism works as follows: A single instance of a functional interface implementation is
created within the GameLevel and made accessible in the ActorManager (through the interface
designed to be exposed to game actors). This implementation tests if a token passed to it is
not null and equal to itself (in other words, it checks whether the caller of the method has
access to its own instance). Because there is only one instance per game level and the player
is not allowed access to it, it is sufficiently robust, if vulnerable to reflection, although
that will probably be dealt with via static analysis of player code, or maybe just its imports.

Also modified existing code to use this new mechanism.

TODO: Split scores per player.