Minor proxies and interface rework
We found a workaround to the previous commit that introduced a "privileged" functionality
subset for the GameWorld object. Instead of implementing the proxy methods for the safe
method subset to expose to the player directly in the player proxy class returns a proxy
object for the game world instead of the actual game world object.

This proxy is implemented as an inner class within the player proxy object.

Also some minor interface changes and code cleanup.