Minor Wall and Enemy related bugfixes
Upon resetting the game updater classes, some member variables were not reset
to their default values, causing strange behavior (actors not generated, etc.)

Also, with respect to the Walls: they are at their current positions left edge
so they unset the neighboring cell to the left of their current position, and
also the neighboring cells neighbor to the right.

Also found and fixed a (presumably) old bug. The wall would sometimes be put
next to an entrance to a tunnel, causing the player to be unable to turn left
and destroy it.