Reworking of player manager in response to previous commit (ea8fc0d)
For reasons similar to those which made us switch from lists to sets in the actor manager
we are also switching the player manager to use sets, as there should be a unique player
instance (additionally, of a unique class as well) there should also be a unique instance
of a pipe object, which said player controls.

Also changed some code in response to this switch from lists to sets.