Refactored game world in a similar way as the actor and player managers
In the previous commits we have switched from lists to sets for storing
references to game actors in the actor manager (ea8fc0d) and for storing
references to pipes of players in the player manager (1da47a2).

Since tunnels should also not have duplicate references in the game world
or in the results of queries performed on it, we have switched from lists
to sets here.

Furthermore, since references to cells that make up a tunnel object should
also not have any duplicates in it or in any query results, we have also
switched to sets for storing them.

Also, code changes in some other classes because of the switch from lists
to sets.