Changing visual display and gameplay and minor bugfixes.
The game screen was slightly changed. Specifically:

- The numerical value that used to display the score for the (single) player
now displays the remaining pieces of treasure to be collected during the level.
- The visual representation of each players weapons bullet queue is positioned
next to the entrance assigned to that particular player.
- After the game has finished, a black screen is shown displaying the scores.
The scores displayed are both scores per level and player class and the total
game score of all players.

Also, a few bugfixes, notably:
- In a previous commit (1da47a2) we have determined that it doesn't make sense
for 2 instances of the same Player class running the same algorithm to compete
against each other. However, since an assignment could also involve implementing
a single algorithm which is designed to cooperate with other instances of itself
we have decided to switch to a different model.

In this model, the player manager keeps scores per player instance, but our game
manager, which now archives the scores that are displayed at the end of the game
stores scores for each level and for the entire game per distinct Player classes.
If we were to use 2 instances of the same concrete Player class in the same level,
the score for that particular class both for the level and within the total scores
would be the combined score of all instances of that class in that level.