Refactored interfaces to make student API a little smaller.
Although the authentication mechanism for certain methods which are exposed to students
in the API, such as Actor.act() should prevent cheats, a problem is that it creates bloat
in it with methods the student can't even use anyway. To address this potential usability
issue, we have decided to apply the principle from the game backend classes of putting
these methods into interfaces representing higher access privileges. Although downcasting
is still possible, the student still couldn't use methods that this would expose. This
should create a view of the API that by default is not populated with unusable methods.

The specific changes so far:

1.) Actor interface no logner contains the act() method, it has been moved to ActorPrivileged.
    The method is only visible to the actor manager and at the time of an actor instantiation
    in any class where this instatiation takes place.
2.) ActorManager exposes only the base Actor interface in all of its method signatures except
    registerActor() where an ActorPrivileged type of parameter is expected.
3.) Pipe had all methods that would probably be used by the player extracted into an interface
    PipeBasic which it implements (getController() could also be put there, but we're unsure
    whether any player implementation would ever even use it).

Also changes to classes implementing these interfaces.

Also: Modified getActorsBySearchCriteria() method of actor manager to return all game actors
if a null value is passed as the criteria parameter.