Modified player manager to return 2 different views of pipes.
Similar to commit c97882e, where we extracted safe methods the Player
can use into an interface, the player manager was changed to return
to players a "safe view" of the set of all pipes in the game.

The regular view exposing methods such as setHealth() still exists and
is available to game actors and backend classes through a new interface
PlayerManagerUpdatable.

Also changes in code because of this.