Actor intersection mechanism changed, probably better now.
Previously, there was a major problem with detecting intersecting actors
simply by checking if the references to their current position (TunnelCell)
referred to the same object, because of which sometimes for example an enemy
was not "eaten" by the pipe head, or similar issues. Now, each actor carries
with it a rectangle object of our own implementation which can check if it
overlaps with another rectangle that is passed as argument.

Each rectangle can return both width, height, as well as 2 sets of coordinates.
One of them are the coordinates of its center, which are the now the "official"
X, Y coordinates of the actor. The other are coordinates used by the rendering
library to correctly place the sprite representing the actor on the screen. We
had to slightly modify the wall sprite though, since it is a special case.

The rectangle object implementation is coordinate system and thus rendering
library dependent, but its implemented interface is not. Since this feature
could be used even for tunnel cells, we have split the implementation and
interfaces into two, with ActorRectangle being the subclass of Rectangle.
ActorRectangle allows translating (basically moving) a rectangle, and also
it is the place where the current position of the actor is kept. We have
moved the core code from AbstractMovable actor here. However, because we've
also modified the ActorBasic interface to return such an object, we've made
a private inner class which represents an unmodifiable view of the actual
ActorRectangle object of the actor (it just forbids translate() from having
any effect when called from outside a subclass).

Also changes to some other code because of this.