Modified gradle files to create a distributable project version.
The schema works thusly: first, an executable JAR file (currently a fat JAR)
with the desktop launcher is created. This JAR excludes all implementations
of the player except the AbstractPlayer class. Afterwards, this JAR file is
packaged together with the contents of the assets folder (except the 'GIMP'
subdirectory) to create a ZIP archive in the projects root directory.

A student thus downloads this ZIP and extracts its content into a directory
where they set up a new java project (in IntelliJ IDEA or any other IDE),
and adds this JAR to the project as a dependency or library. Afterwards,
they set the main class of their project to the JARs DesktopLauncher class
and programs the solution, in other words, a new Player class in the Player
classes package (szakacs.kpi.fei.tuke.player).

Because the player classes used are configured in XML files that are not put
inside the jar but packaged with it into the ZIP, one only needs to modify
some lines in those XML files to use and test their new solution.