Modified actor manager interface.
Actor manager can now return sets of actors of the same ActorType
and sets of actors occupying a common position.

These changes are part of a series of commits made to create easier (and
also faster, in terms of performance) information gathering by players
and to create better actor collision logic.